requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
google-generativeai==0.8.3
google-cloud-secret-manager
